精简stt下各模型私有及共有代码
