--- hlp/stt/utils/record.py
@@ -1,47 +1,0 @@
-import wave
-import pyaudio
-
-
-def record(record_path, record_duration):
-    """从麦克风录音声音道文件
-
-    :param record_path: 声音文件保存路径
-    :param record_duration: 录制时间，秒
-    :return: 无
-    """
-    CHUNK = 1024
-    FORMAT = pyaudio.paInt16
-    CHANNELS = 1  # 声道数
-    RATE = 16000  # 采样率
-    RECORD_SECONDS = record_duration
-    WAVE_OUTPUT_FILENAME = record_path
-    p = pyaudio.PyAudio()
-
-    stream = p.open(format=FORMAT,
-                    channels=CHANNELS,
-                    rate=RATE,
-                    input=True,
-                    frames_per_buffer=CHUNK)
-    frames = []
-    print("开始录音：请在%d秒内输入语音:" % RECORD_SECONDS)
-    for i in range(1, int(RATE / CHUNK * RECORD_SECONDS) + 1):
-        data = stream.read(CHUNK)
-        frames.append(data)
-        if (i % (RATE / CHUNK)) == 0:
-            print('\r%s%d%s' % ("剩余", int(RECORD_SECONDS - (i // (RATE / CHUNK))), "秒"), end="")
-    print("\r录音结束\n")
-
-    stream.stop_stream()
-    stream.close()
-    p.terminate()
-
-    wf = wave.open(WAVE_OUTPUT_FILENAME, 'wb')
-    wf.setnchannels(CHANNELS)
-    wf.setsampwidth(p.get_sample_size(FORMAT))
-    wf.setframerate(RATE)
-    wf.writeframes(b''.join(frames))
-    wf.close()
-
-
-if __name__ == "__main__":
-    record('temp.wav', 5)
